+++ src/main/java/webserver/WebApplicationServer.java
@@ -11,7 +11,7 @@ import api.UserController;
 import webserver.handler.Handler;
 import webserver.handler.HandlerComposite;
 import webserver.request.HttpMethod;
-import webserver.request.HttpRequestConverter;
+import webserver.handler.HttpRequestConverter;
 import webserver.response.HttpResponseRenderer;
 
 public class WebApplicationServer {
