+++ src/main/java/webserver/handler/HttpRequestConverter.java
@@ -1,4 +1,4 @@
-package webserver.request;
+package webserver.handler;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,6 +6,8 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.List;
+
+import webserver.request.HttpRequest;
 
 public class HttpRequestConverter {
